requirements.txt
numpy==1.24.4
pandas==2.0.3
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
locust==2.18.1
locust-plugins
locust-plugins[dashboards]
locust-plugins[listeners]
pyquery
requests-kerberos==0.12.0
pyhive
python-dotenv